+++ CHANGELOG.md
@@ -10,8 +10,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - Support for [Automation API](https://docs.courier.com/reference/automation-api) by @tk26
-  - `POST /automations/invoke` with `client.automations.invokeAdHocAutomation` method
-  - `POST /automations/{template_id}/invoke` with `client.automations.invokeAutomationTemplate` method
+  - `POST /automations/invoke` with `client.automations.invoke_ad_hoc_automation` method
+  - `POST /automations/{template_id}/invoke` with `client.automations.invoke_automation_template` method
 
 ## [v1.7.0] - 2020-12-18